f3

+++ SpartanWrath/src/main/java/com/spartanwrath/controller/AdminController.java
@@ -1,17 +1,24 @@
 package com.spartanwrath.controller;
 
+import com.spartanwrath.exceptions.NoSuchClass;
 import com.spartanwrath.model.CombatClass;
+import com.spartanwrath.model.Membership;
 import com.spartanwrath.service.CombatClassService;
 import com.spartanwrath.service.MembershipService;
 import com.spartanwrath.service.UserService;
+import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 
 import java.io.IOException;
+import java.security.Principal;
 import java.util.List;
 import java.util.Optional;
 
@@ -23,6 +30,7 @@ public class AdminController {
     MembershipService membershipService;
     @Autowired
     CombatClassService combatClassService;
+
 
     @GetMapping("/Admin")
     public String showAdmin(){
@@ -59,8 +67,16 @@ public class AdminController {
     }
 
     @GetMapping("/Admin/combatclass/{id}/delete")
-    public String deleteCombatClass(@PathVariable long id) {
-
+    public String deleteCombatClass(@PathVariable long id) throws NoSuchClass {
+        Optional<CombatClass> combatClass = combatClassService.findById(id);
+        if (combatClass.isEmpty()){
+            throw new NoSuchClass();
+        }
+        List<Membership> memberships = combatClass.get().getMemberships();
+        for (Membership membership : memberships) {
+            membership.setCombatClass(null);
+            membershipService.save(membership);
+        }
         combatClassService.delete(id);
 
         return "redirect:/Admin/combatclass";
